requirements.txt
Flask==2.3.2
joblib==1.2.0
numpy==1.24.2
pandas==2.0.0
pyaml==23.5.9
scikit-learn==1.2.2
tqdm==4.65.0
typer==0.9.0